New perdaix_coordinates.conf file: Fixes wrong counter positions in TRD.
Found out that the TRD positions and counter positions had different x coordinates --> stereo angle != 0
This lead to wrong chi2 values and was visible in the event display were TRD hits "moved" although they shouldn't (thanks David for the observation).
Andreas found a bug in the perdaixGeometry program in the CVS perdaix10 software and fixed it. This new config file represents the output of perdaixGeometry
with the fix.