Removed clearClusters and deleteClusters from Setup since DetectorElements are not thread-safe.
Dynamically generated clusters (i.e. _NOT_ read from a file) should now be deleted by qDeleteAll(clusters).
I commented this in tools/analysis though, because handling of non-cluster files is broken anyway right now.
It's probably not necessary at all to handle non-cluster files in tools/analysis, so it might stay the way it is.
In the eventdisplay it's useful to be able to open raw-data files without runnign preAnalysis previously, though.
Eventdisplay won't be multithreaded, so there is no problem with a qDeleteAll() there.