Big Performance boost in TRDSpectrumPlot! Profiling showed that TRDSpectrumPlot used up 26% of processing time, reduced to 9% now.
- Removed an unnecessary QVector which was allocated on the stack and dynamically expanded by push_backs. That's slow!
- Replaced all foreach macros

Although we pass a reference to a QVector/std::vector as the second argument, Qt copies the vector!
Passus from Qt Documentation:

"Qt automatically takes a copy of the container when it enters a foreach loop. If you modify the container as you are iterating,
that won't affect the loop. (If you do not modify the container, the copy still takes place, but thanks to implicit sharing copying a container is very fast.)"

It turns out that this caused quite a big slowdown in TRDSpectrumPlot because foreach was called millions of times. Now using iterators instead, which avoid copying.