Fixed regression in TRD plots: They now use only the hits on tracks again.
In commit 2dcbbb2fd17993b7e0250a7c094686cd03f05dca I changed the content of the first argument of
processEvent(). Before that commit it contained only the hits on the track, after the commit it
contained all clusters in the event. This is to be distinguished from event->hits(), which _could_
in principal contain raw hits (i.e. non clusters) when analysis is run on raw data.