AnalysisProcessor: Enable mutex locking for processEvent.
If you want to use it, your EventDestination needs to use "setNeedsLocking(true)". This has been done for all AnalysisPlots automatically, but not for others.
Although Roman and I were already convinced that TH1 is not thread-safe, we never encountered any errors without mutex lockers.
Now that we found some (especially in CutStatisticsPlot), adding a mutex turned out to be a necessity.
The whole processEvent function is locked, this is safer than just locking the "Fill" step. The performance drawback is there, but we can live
with it because other threads can just process other Destinations first (we have 825 of them in the main analysis). The patch adds a list of "postponed"
destinations in that case (the ones that were locked at the first try) and makes the analysis processor loop over them afterwards.