Added a tool to diff postAnalysis files.
The tool can be used to test changes to the analysis core.
The changes should not influence the result plots.

Already I found that the multithreading stack must have a bug
because the results of a dual core analysis always differs.
Analyzing on one core produces the same plots.

The software reads histograms out of two postAnalysis files
and compares their bin contents. So far 1D and 2D histograms
are implemented. Graphs will follow.