Massively simplified track fitting (mathematically it should still be the same).
It's much better to solve all equations in their local (u,v,z) frame. This
reduces the number of required rotations significantly. I now compare only the u
coordinate. The v coordinate is not measured. The equation in the local frame is
then:

cos(alpha)*(x0+k*slopeX) - sin(alpha)*(y0+k*slopeY)
 = u = cos(alpha)*fX - sin(alpha)*fY

Where x0, y0, slopeX, slopeY are the track parameters in (x,y,z) and alpha is
the angle of the local fiber layer with respect to the y-axis. fX and fY are the
x and y coordinates of the hit in this layer.

- We don't need any tangens or cotangens anywhere - i.e. no removal of
  singularities required.
- No more pivoting ("useTangens") and combining of x and y equations by-hand
  required.
- No more calculation of strange covariance matrix elements required (we can
  directly use sigmaU).

Curious artifact still left over: I need to rotate by -angle to make things
work, which I don't understand at the moment.

Possible Explanations:

1. The rotation matrices I use is actually rotation in mathematical negative
   sense (unlikely).
2. All our tracker layers in the coordinates config file have exactly the wrong
   sign of their rotation. This is not impossible because you wouldn't notice
   it from the hit coordinates themselves (apriori = 0.5*(pos + counterPos)
   i.e. on the z-axis). Instead you would only notice it in the reconstruction,
   which is what I'm seeing so it might be possible.