Introduced new Corrections step: postTOFCorrections.
This is because we found that particle->timeOfFlight()->beta() was used before
tof->calculateTimes(track) was called. Therefore the time information in the tof
object was invalid/from the previous event.

In f8ac42f7535bdfb0b683b1026a268f1e7c9af04b a reset function was added. But the
order still had to be changed.

One small drawback of the new approach is that the photon travel time will not
use the best guess for the track anymore, but I found that the differences
between track before multiple scattering corrections and after are very small,
so maybe we can live with this.