Changed cluster processing of sub hits.
The cluster now stores the detector id of the strip with the highest amplitude.
This affects only the "channel" part of the id (i.e. for the tracker the last 5
bits.). The "element" part was already set in the old clusters. The y coordinate
of the cluster is now also computed from the weighted mean of the strips.