Fixed ClusterShapePlot and made it faster as well.
I think the code was broken, because it checked all hits in the event against
the maximum strip, instead of checking only the hits belonging to the same
cluster. Probably it ended up in the under/overflow of the histogram, but it's
better this way. Also uses the new channel() bits in cluster to determine the
channel of the maximum. This allows me to drop some really crappy code from
before. Also uses the sortedChannel() function now.