In LimitedEventQueue reserve space for m_bufferSize events. This avoids reallocation of the internal data array.
QQueue inherits QList, which stores pointers to the data in an internal linear
array. In our case it stores the pointers to the events directly, it is thus
more or less equivalent to a QVector. Because appending and prepending changes
the length of the used portion of that array, reallocation has to occur
occasionally. Knowing the maximum size of the queue beforehand, we can prevent
this by reserving appropriate amount of space.