Increasing the storage size of the event queues to 10.000 events.
Although, from the gui it looks like the queues are empty and the reader thread
has its hands full keeping up the work, profiling shows that this is not
so. Please report if you're having memory issues now. We can arrange a better
value, or revert this change if need be.