Introduced a passes(event) and a passes(hits, particle) function in CutFilter and Cuts.
I liked David's idea, but I was not fond of the special place the cherenkov cut
had in his commit. Things should try to make use of the framework we have,
i.e. all cuts go in the CutFilter. The modified version gives greater
flexibility in the sense that every cut can now apply pre-reconstruction cuts
and post-reconstruction cuts if they want to.